+++ src/RSoft.Auth.Web.Api/Controllers/v1_0/RolesController.cs
@@ -24,6 +24,7 @@ namespace RSoft.Auth.Web.Api.Controllers.v1_0
     /// </summary>
     [ApiVersion("1.0")]
     [Route("api/v{version:apiVersion}/[controller]")]
+    //[Route("api/v{version:apiVersion}/{culture:culture}/[controller]")]
     [ApiController]
     [Authorize(Policy = PolicyNames.OnlyThisApplication)]
     public class RolesController : ApiCrudBaseController<Guid, RoleDto, RoleRequest, RoleResponse>