+++ src/RSoft.Auth.Web.Api/Language/Resource.cs
@@ -1,0 +1,10 @@
+﻿namespace RSoft.Auth.Web.Api.Language
+{
+
+    /// <summary>
+    /// Resource languager holder
+    /// </summary>
+    public class Resource
+    {
+    }
+}