+++ src/RSoft.Auth.Web.Api/Model/Request/v1_0/PasswordRequest.cs
@@ -12,8 +12,8 @@ namespace RSoft.Auth.Web.Api.Model.Request.v1_0
         /// <summary>
         /// User login (email
         /// </summary>
-        [Required(ErrorMessage = "Login is required")]
-        [MaxLength(254, ErrorMessage = "Login must contain a maximum of 254 characters")]
+        [Required(ErrorMessage = "LOGIN_REQUIRED")]
+        [MaxLength(254, ErrorMessage = "LOGIN_MAX_SIZE")]
         public string Login { get; set; }
 
     }